requirements_test.txt
pytest-homeassistant-custom-component
filelock
pytest-cov